functions/pyproject.toml
[tool.poetry]
name = "functions"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "^1.14.0"
fastapi = "^0.96.0"
fastapi-versioning = "^0.10.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
authlib = "^1.2.1"
httpx = "^0.24.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pylint = "^2.17.4"
uvicorn = "^0.22.0"
pydantic-to-typescript = "^1.0.10"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
